Adding new combined code

+++ BoostedTopAnalysis/python/Type11SelectionParams_cfi.py
@@ -5,7 +5,7 @@ from PhysicsTools.SelectorUtils.pfJetIDSelector_cfi import pfJetIDSelector
 Type11QCDEstimationParams = cms.PSet( 
   Type11Selection = cms.PSet(
       pfJetIDParams = cms.PSet( pfJetIDSelector.clone() ),
-    caTopJetPtMin = cms.double(250),
+    caTopJetPtMin = cms.double(30),
     caTopJetEtaCut = cms.double(2.4),
     patJetCollectionInputTag = cms.InputTag( "selectedPatJetsPFlow" ),
     caTopJetCollectionInputTag = cms.InputTag( "selectedPatJetsCATopTagPF" )
